#144 refactor: 리포트 화면은 화면 전환이 없는 단일 화면이므로 ReportCoordinator를 deprecated 시키고 ReportViewController를 직접 push하도록 변경
