+++ Projects/Features/FeatureHome/FeatureHomeCoordinator/Project.swift
@@ -19,6 +19,7 @@ let project = Project.makeModule(
         .Project.module(.Features(.FriendList, .CoordinatorInterface)).dependency,
         .Project.module(.Features(.Home, .CoordinatorInterface)).dependency,
         .Project.module(.Features(.Home, .Presentation)).dependency,
+        .Project.module(.Features(.Report, .Presentation)).dependency,
         .Project.module(.Features(.Setting, .Coordinator)).dependency,
         .Project.module(.Features(.Setting, .CoordinatorInterface)).dependency,
         .Project.module(.Features(.StoolLog, .Coordinator)).dependency,
